Move `FormatOptions` into a separate file
