+++ src/lib.rs
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use anyhow::{anyhow, bail, Error, Result};
+use formatter::FormatOptions;
 use grammar::{
     is_turtle_decimal, is_turtle_double, is_turtle_integer, NodeKind, RootContext, StringDecoder,
 };
@@ -11,56 +12,8 @@ use std::collections::HashMap;
 use std::fmt::Write;
 use tree_sitter::{Language, Node};
 
+pub mod formatter;
 mod grammar;
-
-pub struct FormatOptions {
-    /// Number of spaces used for one level of indentation
-    pub indentation: usize,
-    /// Whether to sort subjects, predicates and objects,
-    /// including within blank-nodes
-    pub sort_terms: bool,
-    /// Enables inserting new-lines before the following:
-    /// - a subjects finalizing dot
-    /// - the first predicate of a subject
-    /// - the first object within one subject-predicate pair
-    /// - each objects within one subject-predicate pair
-    /// - each collection item;
-    ///   see <https://www.w3.org/TR/rdf12-turtle/#collections>
-    /// - each predicate within a blank-node
-    pub new_lines_for_easy_diff: bool,
-    /// Whether to move a single/lone object
-    /// (within one subject-predicate pair) onto a new line,
-    /// or to keep it on the same line as the predicate.
-    pub single_object_on_new_line: bool,
-    /// Whether to cleanup/unify empty lines used as dividers.
-    /// This ensures that there is exactly one empty line
-    /// before and after each subject,
-    /// and that there is none anywhere else.
-    pub cleanup_dividing_empty_lines: bool,
-    /// Whether to force-write the output,
-    /// even if potential issues with the formatting have been detected.
-    pub force: bool,
-}
-
-impl Default for FormatOptions {
-    fn default() -> Self {
-        Self {
-            indentation: 4,
-            sort_terms: false,
-            new_lines_for_easy_diff: false,
-            single_object_on_new_line: false,
-            cleanup_dividing_empty_lines: false,
-            force: false,
-        }
-    }
-}
-
-impl FormatOptions {
-    #[must_use]
-    pub const fn includes_sorting(&self) -> bool {
-        self.sort_terms
-    }
-}
 
 struct PrefixedResource {
     /// The prefix; e.g. `"xsd"` or `"schema"`.
